Refactory to tableFilterUtil/tableFilterTypes
