+++ src/modules/ce/components/client/ClientOpportunitiesTable.tsx
@@ -5,8 +5,8 @@ import { ColumnDef } from '@/components/elements/table/types';
 import useClientDashboardContext from '@/modules/client/hooks/useClientDashboardContext';
 import GenericTableWithData from '@/modules/dataFetching/components/GenericTableWithData';
 import ProjectTypeChip from '@/modules/hmis/components/ProjectTypeChip';
-import { useFilters } from '@/modules/hmis/filterUtil';
 import { clientBriefName } from '@/modules/hmis/hmisUtil';
+import { useFilters } from '@/modules/hmis/tableFilterUtil';
 import { ProjectDashboardRoutes } from '@/routes/routes';
 import {
   ClientCeOpportunitySummaryFieldsFragment,
