+++ src/modules/enrollment/components/EnrollmentAssessmentsTable.tsx
@@ -5,8 +5,8 @@ import {
   generateAssessmentPath,
 } from '@/modules/assessments/util';
 import GenericTableWithData from '@/modules/dataFetching/components/GenericTableWithData';
-import { useFilters } from '@/modules/hmis/filterUtil';
 import { assessmentDescription } from '@/modules/hmis/hmisUtil';
+import { useFilters } from '@/modules/hmis/tableFilterUtil';
 import {
   AssessmentFieldsFragment,
   GetEnrollmentAssessmentsDocument,
