+++ src/modules/enrollment/components/HouseholdAssessmentsTable.tsx
@@ -6,11 +6,11 @@ import {
   generateAssessmentPath,
 } from '@/modules/assessments/util';
 import GenericTableWithData from '@/modules/dataFetching/components/GenericTableWithData';
-import { useFilters } from '@/modules/hmis/filterUtil';
 import {
   assessmentDescription,
   clientBriefName,
 } from '@/modules/hmis/hmisUtil';
+import { useFilters } from '@/modules/hmis/tableFilterUtil';
 import {
   GetHouseholdAssessmentsDocument,
   GetHouseholdAssessmentsQuery,
