+++ src/modules/projects/components/ProjectAssessments.tsx
@@ -12,11 +12,11 @@ import {
 } from '@/modules/assessments/util';
 import GenericTableWithData from '@/modules/dataFetching/components/GenericTableWithData';
 import { WITH_ENROLLMENT_COLUMNS } from '@/modules/enrollment/columns/enrollmentColumns';
-import { useFilters } from '@/modules/hmis/filterUtil';
 import {
   assessmentDescription,
   clientBriefName,
 } from '@/modules/hmis/hmisUtil';
+import { useFilters } from '@/modules/hmis/tableFilterUtil';
 import {
   AssessmentSortOption,
   GetProjectAssessmentsDocument,
