+++ src/modules/projects/components/tables/ProjectExternalSubmissionsTable.tsx
@@ -7,7 +7,7 @@ import theme from '@/config/theme';
 import GenericTableWithData from '@/modules/dataFetching/components/GenericTableWithData';
 import ApolloErrorAlert from '@/modules/errors/components/ApolloErrorAlert';
 import RelativeDateTableCellContents from '@/modules/hmis/components/RelativeDateTableCellContents';
-import { useFilters } from '@/modules/hmis/filterUtil';
+import { useFilters } from '@/modules/hmis/tableFilterUtil';
 import ExternalSubmissionsViewModal from '@/modules/projects/components/ExternalSubmissionsViewModal';
 import RefreshExternalSubmissionsButton from '@/modules/projects/components/RefreshExternalSubmissionsButton';
 import {
