+++ src/modules/services/components/ProjectServicesTable.tsx
@@ -5,8 +5,8 @@ import ClientName from '@/modules/client/components/ClientName';
 import GenericTableWithData from '@/modules/dataFetching/components/GenericTableWithData';
 
 import { WITH_ENROLLMENT_COLUMNS } from '@/modules/enrollment/columns/enrollmentColumns';
-import { useFilters } from '@/modules/hmis/filterUtil';
 import { clientBriefName, parseAndFormatDate } from '@/modules/hmis/hmisUtil';
+import { useFilters } from '@/modules/hmis/tableFilterUtil';
 import {
   getServiceTypeForDisplay,
   SERVICE_BASIC_COLUMNS,
